requirements.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements.in
#
beautifulsoup4==4.10.0
    # via -r requirements.in
camelot-py[cv]==0.10.1
    # via -r requirements.in
certifi==2021.10.8
    # via requests
cffi==1.15.0
    # via cryptography
chardet==4.0.0
    # via
    #   camelot-py
    #   pdfminer.six
charset-normalizer==2.0.9
    # via requests
click==8.0.3
    # via
    #   camelot-py
    #   pdftopng
cryptography==36.0.0
    # via pdfminer.six
et-xmlfile==1.1.0
    # via openpyxl
ghostscript==0.7
    # via camelot-py
idna==3.3
    # via requests
numpy==1.21.4
    # via
    #   camelot-py
    #   opencv-contrib-python-headless
    #   opencv-python
    #   pandas
opencv-contrib-python-headless==4.5.4.60
    # via -r requirements.in
opencv-python==4.5.4.60
    # via camelot-py
openpyxl==3.0.9
    # via camelot-py
pandas==1.3.4
    # via camelot-py
pdfminer.six==20211012
    # via camelot-py
pdftopng==0.2.3
    # via camelot-py
pycparser==2.21
    # via cffi
pypdf2==1.26.0
    # via camelot-py
python-dateutil==2.8.2
    # via pandas
pytz==2021.3
    # via pandas
requests==2.26.0
    # via -r requirements.in
six==1.16.0
    # via python-dateutil
soupsieve==2.3.1
    # via beautifulsoup4
tabulate==0.8.9
    # via camelot-py
urllib3==1.26.7
    # via requests

# The following packages are considered to be unsafe in a requirements file:
# setuptools
